+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/types/Either.kt
@@ -2,61 +2,119 @@
 
 package expo.modules.kotlin.types
 
+import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.apifeatures.EitherType
+import java.lang.ref.WeakReference
 import kotlin.reflect.KClass
+import kotlin.reflect.KType
+
+sealed class DeferredValue
+
+object IncompatibleValue : DeferredValue()
+
+class UnconvertedValue(
+  private val unconvertedValue: Any,
+  private val typeConverter: TypeConverter<*>,
+  context: AppContext?
+) : DeferredValue() {
+  private val contextHolder = WeakReference(context)
+
+  private var convertedValue: Any? = null
+
+  fun getConvertedValue(): Any {
+    if (convertedValue == null) {
+      convertedValue = typeConverter.convert(unconvertedValue, contextHolder.get())
+    }
+
+    return convertedValue!!
+  }
+}
+
+data class ConvertedValue(val convertedValue: Any) : DeferredValue()
 
 @EitherType
 open class Either<FirstType : Any, SecondType : Any>(
-  @PublishedApi
-  internal val value: Any
+  private val bareValue: Any,
+  private val deferredValue: MutableList<DeferredValue>,
+  private val types: List<KType>
+
 ) {
-  @JvmName("isFirstType")
-  fun `is`(type: KClass<FirstType>): Boolean {
-    return type.isInstance(value)
+  internal fun `is`(index: Int): Boolean {
+    return when (val value = deferredValue[index]) {
+      is ConvertedValue -> true
+      IncompatibleValue -> false
+      is UnconvertedValue -> {
+        try {
+          deferredValue[index] = ConvertedValue(value.getConvertedValue())
+          true
+        } catch (e: Throwable) {
+          deferredValue[index] = IncompatibleValue
+          false
+        }
+      }
+    }
   }
+
+  internal fun get(index: Int): Any {
+    return when (val value = deferredValue[index]) {
+      is ConvertedValue -> value.convertedValue
+      IncompatibleValue -> throw TypeCastException("Cannot cast '$bareValue' to '${types[index]}'")
+      is UnconvertedValue -> {
+        try {
+          val convertedValue = value.getConvertedValue()
+          deferredValue[index] = ConvertedValue(convertedValue)
+          convertedValue
+        } catch (e: Throwable) {
+          deferredValue[index] = IncompatibleValue
+          throw TypeCastException("Cannot cast '$bareValue' to '${types[index]}'")
+        }
+      }
+    }
+  }
+
+  @JvmName("isFirstType")
+  fun `is`(type: KClass<FirstType>): Boolean = `is`(0)
 
   @JvmName("isSecondType")
-  fun `is`(type: KClass<SecondType>): Boolean {
-    return type.isInstance(value)
-  }
+  fun `is`(type: KClass<SecondType>): Boolean = `is`(1)
 
   @JvmName("getFirstType")
-  fun get(type: KClass<FirstType>) = value as FirstType
+  fun get(type: KClass<FirstType>): FirstType = get(0) as FirstType
 
   @JvmName("getSecondType")
-  fun get(type: KClass<SecondType>) = value as SecondType
+  fun get(type: KClass<SecondType>): SecondType = get(1) as SecondType
 
-  fun first() = value as FirstType
+  fun first(): FirstType = get(0) as FirstType
 
-  fun second() = value as SecondType
+  fun second(): SecondType = get(1) as SecondType
 }
 
 @EitherType
 open class EitherOfThree<FirstType : Any, SecondType : Any, ThirdType : Any>(
-  value: Any
-) : Either<FirstType, SecondType>(value) {
+  bareValue: Any,
+  deferredValue: MutableList<DeferredValue>,
+  types: List<KType>
+) : Either<FirstType, SecondType>(bareValue, deferredValue, types) {
   @JvmName("isThirdType")
-  fun `is`(type: KClass<ThirdType>): Boolean {
-    return type.isInstance(value)
-  }
+  fun `is`(type: KClass<ThirdType>): Boolean = `is`(2)
 
   @JvmName("getThirdType")
-  fun get(type: KClass<ThirdType>) = value as ThirdType
+  fun get(type: KClass<ThirdType>) = get(3) as ThirdType
 
-  fun third() = value as ThirdType
+  fun third(): ThirdType = get(3) as ThirdType
 }
 
 @EitherType
 class EitherOfFour<FirstType : Any, SecondType : Any, ThirdType : Any, FourthType : Any>(
-  value: Any
-) : EitherOfThree<FirstType, SecondType, ThirdType>(value) {
+  bareValue: Any,
+  deferredValue: MutableList<DeferredValue>,
+  types: List<KType>
+) : EitherOfThree<FirstType, SecondType, ThirdType>(bareValue, deferredValue, types) {
   @JvmName("isFourthType")
-  fun `is`(type: KClass<FourthType>): Boolean {
-    return type.isInstance(value)
-  }
+  fun `is`(type: KClass<FourthType>): Boolean = `is`(3)
 
   @JvmName("getFourthType")
-  fun get(type: KClass<FourthType>) = value as FourthType
+  fun get(type: KClass<FourthType>): FourthType = get(3) as FourthType
 
-  fun fourth() = value as FourthType
+  fun fourth(): FourthType = get(3) as FourthType
 }